Fix loading of iOS 4 device playlists
This fixes a few problems related to the loading of on-the-go playlists on iOS 4 devices.

(The main problem was that they don't contain the `dbTimestampMacOsDate` property that nano 7G on-the-go playlists have.)

It also changes the logic so that if an on-the-go playlist cannot be loaded, a warning is logged to the console rather than aborting the entire load of the device's database.